Added Hero model. Renamed 'common' to 'shared'. Minor refactor to heroes.mock.ts
