+++ src/app/heroes/heroDetail.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { RouteParams } from '@angular/router-deprecated';
 
-import { HeroService } from '../common';
+import { Hero, HeroService } from '../shared';
 
 @Component({
   selector: 'ngp-hero-detail',
@@ -9,13 +9,13 @@ import { HeroService } from '../common';
   styles: [require('./heroDetail.component.scss')],
 })
 export class HeroDetailComponent implements OnInit {
-  hero = undefined;
+  hero: Hero;
 
   constructor(private heroService: HeroService, private routeParams: RouteParams) {
   }
 
   ngOnInit() {
-    const id = +this.routeParams.get('id');
+    let id = +this.routeParams.get('id');
     this.heroService.getHero(id).then(hero => (this.hero = hero));
   }
