+++ src/app/heroes/heroes.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router-deprecated';
 
-import { HeroService } from '../common';
+import { Hero, HeroService } from '../shared';
 
 @Component({
   selector: 'ngp-heroes',
@@ -9,8 +9,8 @@ import { HeroService } from '../common';
   styles: [require('./heroes.component.scss')],
 })
 export class HeroesComponent {
-  heroes: any[];
-  selectedHero: any;
+  heroes: Hero[];
+  selectedHero: Hero;
 
   constructor(private router: Router, private heroService: HeroService) {
   }
