+++ src/app/shared/hero.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 
-import HEROES from './heroes.mock';
+import { Hero } from './';
+import { HEROES } from './heroes.mock';
 
 @Injectable()
 export class HeroService {
@@ -9,12 +10,12 @@ export class HeroService {
   }
 
   getHeroesSlowly() {
-    return new Promise<any[]>(resolve =>
+    return new Promise<Hero[]>(resolve =>
       setTimeout(() => resolve(HEROES), 5000)
     );
   }
 
   getHero(id: number) {
-    return this.getHeroes().then((heroes: any[]) => heroes.filter(hero => hero.id === id)[0]);
+    return this.getHeroes().then(heroes => heroes.filter(hero => hero.id === id)[0]);
   }
 }
