+++ src/app/shared/hero.ts
@@ -1,0 +1,4 @@
+export class Hero {
+  id: number;
+  name: string;
+}
