+++ src/app/shared/heroes.mock.ts
@@ -1,4 +1,6 @@
-export default [
+import { Hero } from './';
+
+export var HEROES: Hero[] = [
   { id: 11, name: 'Mr. Nice' },
   { id: 12, name: 'Narco' },
   { id: 13, name: 'Bombasto' },
