+++ src/app/shared/index.ts
@@ -1,1 +1,2 @@
+export * from './hero';
 export * from './hero.service';
